+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/eslint-config
 
+## 2.0.1
+
+### Patch Changes
+
+- 144cd19: Upgrade all JavaScript dependencies
+
 ## 2.0.0
 
 ### Major Changes